+++ Assets/OkapiKit/Scripts/Actions/ActionFlashV2.cs
@@ -5,6 +5,9 @@ namespace OkapiKit
     [AddComponentMenu("Okapi/Action/Flash")]
     public class ActionFlashV2 : ActionEffect
     {
+        public enum Mode { ColorFlash, InvertColor, SmoothInvert };
+
+        [SerializeField] private Mode       mode = Mode.ColorFlash;
         [SerializeField] private Gradient   color;
         [SerializeField] private float      duration = 1.0f;
 
@@ -20,18 +23,59 @@ namespace OkapiKit
                 return;
             }
 
-            target.FlashColor(duration, color);
+            switch (mode)
+            {
+                case Mode.ColorFlash:
+                    target.FlashColor(duration, color);
+                    break;
+                case Mode.InvertColor:
+                    target.FlashInvert(duration);
+                    break;
+                case Mode.SmoothInvert:
+                    target.SmoothInvert(duration);
+                    break;
+                default:
+                    break;
+            }
         }
 
         public override string GetRawDescription(string ident, GameObject refObject)
         {
             string desc = GetPreconditionsString(gameObject);
+
+            switch (mode)
+            {
+                case Mode.ColorFlash:
+                    desc += $"flashes renderer {targetRenderer.GetRawDescription("target", gameObject)} with a color for {duration} seconds";
+                    break;
+                case Mode.InvertColor:
+                    desc += $"flashes renderer {targetRenderer.GetRawDescription("target", gameObject)}, inverting the colors for {duration} seconds";
+                    break;
+                case Mode.SmoothInvert:
+                    desc += $"smoothly inverts colors in and out, on renderer {targetRenderer.GetRawDescription("target", gameObject)}, during {duration} seconds";
+                    break;
+                default:
+                    break;
+            }
             
-            desc += $"flashes renderer {targetRenderer.GetRawDescription("target", gameObject)} for {duration} seconds";
 
             return desc;
         }
 
-        public override string GetActionTitle() { return "Flash"; }
+        public override string GetActionTitle() 
+        {
+            switch (mode)
+            {
+                case Mode.ColorFlash:
+                    return "Flash";
+                case Mode.InvertColor:
+                    return "Invert Colors";
+                case Mode.SmoothInvert:
+                    return "Invert Colors";
+                default:
+                    break;
+            }
+            return "[UNDEFINED]"; 
+        }
     }
 }