+++ Assets/OkapiKit/Scripts/Editor/ActionFlashV2Editor.cs
@@ -8,6 +8,7 @@ namespace OkapiKit.Editor
     {
         SerializedProperty color;
         SerializedProperty duration;
+        SerializedProperty mode;
 
         protected override void OnEnable()
         {
@@ -15,11 +16,17 @@ namespace OkapiKit.Editor
 
             color = serializedObject.FindProperty("color");
             duration = serializedObject.FindProperty("duration");
+            mode = serializedObject.FindProperty("mode");
         }
 
         protected override void ActionEffect_OnInspectorGUI()
         {
-            EditorGUILayout.PropertyField(color, new GUIContent("Color", "Color gradient to flash the object.\nIf alpha equals zero, there's no flash at that point, if it's equal to 1, only the flash color will be visible."));
+            EditorGUILayout.PropertyField(mode, new GUIContent("Mode", "Select if you want to flash a specific color, or just invert colors during the flash."));
+            var modeValue = (ActionFlashV2.Mode)mode.enumValueIndex;
+            if (modeValue == ActionFlashV2.Mode.ColorFlash)
+            {
+                EditorGUILayout.PropertyField(color, new GUIContent("Color", "Color gradient to flash the object.\nIf alpha equals zero, there's no flash at that point, if it's equal to 1, only the flash color will be visible."));
+            }
             EditorGUILayout.PropertyField(duration, new GUIContent("Duration", "For how long to flash the object, in seconds."));
         }
     }